A Temperature Based Fan Speed Controller
The main objective of our project is to control the speed of the fan automatically based on the surrounding room temperature without any human intervention and monitor the parameters such as temperature, humidity, and fan speed on the OLED screen. Our other objectives include sending the data of these parameters to a cloud platform and visualize the data in the form of graphs and sending the notifications to our mobile phones using the Blynk application.